+++ Plugins/Object/Object.hpp
@@ -60,6 +60,7 @@ private:
     ArgumentStack GetIsDestroyable          (ArgumentStack&& args);
     ArgumentStack DoSpellImmunity           (ArgumentStack&& args);
     ArgumentStack DoSpellLevelAbsorption    (ArgumentStack&& args);
+    ArgumentStack SetHasInventory           (ArgumentStack&& args);
 
     CNWSObject *object(ArgumentStack& args);
 };